lua_pattern

Function parse

pub fn parse(pattern: impl AsRef<str>) -> Result<Pattern>

Parse the given input string as a Lua pattern.

Returns

This function returns a vector of PatternObjects if parsing was successful, or an Error if the pattern could not be parsed.

Errors

To see the possible errors, have a look at Error.